Redid sliding section
Redid the sliding section of the algorithm. When first coded I thought I
could combine the test to see if the middle to end point line was
intersecting the current body and the test to see if the middle point
was inside a convex with sliding the middle point toward the body.
Sliding the middle point toward the body and sliding away must be in
separate steps because if the middle point starts out in a convex but
would be in an open space after sliding toward the body, it won't
because it is in a convex. So first slide toward the body, then slide
away outside of any convexes.