Fixed some bugs and added a couple tests
Fixed a bug where the path algorithm would exit early because it counted
an unblocked segment that wasn't the end segment as being finished.
Fixed a few other minor things and cleaned up some code.
Added a couple tests, one which is very simple and only tests if the
path can get around one rectangle. The other creates 100 rectangles in a
100x100 space and has the algorithm find a path from the very left side
to the very right side. It also renders the rectangles and path after
the algorithm is finished so the path can be seen.